Attaching random colors
Random Colors and random color methods being attached to ui
fixed issue with pixel in corner being wrong, issue was setting pixel
to -1 when off of the screen.  That was shifting pixel values down by 1
and causing the error.
New issue of brushes smearing color over the x axis, but not the y axis